Renamed PrestaShopVersion to Version
