+++ src/Core/Foundation/Version/Exception/InvalidVersionException.php
@@ -23,7 +23,7 @@
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  * International Registered Trademark & Property of PrestaShop SA
  */
-namespace PrestaShop\PrestaShop\Core\Foundation\PrestaShopVersion\Exception;
+namespace PrestaShop\PrestaShop\Core\Foundation\Version\Exception;
 
 use Exception;
